Removing json fields that would be null from the response.
